+++ src/landingpage/Main.js
@@ -1,9 +1,9 @@
 import React from "react";
-import Navbar from "./Navbar";
+import Navbar from "../components/navbar/Navbar";
 // import Herosection from "./Herosection";
-import Features from "./Features";
+import Features from "../components/features/Features";
 
-import Newsletter from "./Newsletter";
+import Newsletter from "../components/newsletter/Newsletter";
 
 import Hero from "../components/hero/Hero";
 import Howitworks from "../components/howitworks/Howitworks";
